13. DataExpert/Homework/Week 2/Week2-Homework 1/requirements.txt
anthropic
python-dotenv
PyGithub
